Rename library utility.hpp to rocblas_utility.hpp
A header in common/ cannot depend on a header in library/, but
common_host_helpers.hpp does. It only ever compiled because there were
two different headers called utility.hpp.

This is solved by moving the library/ utility.hpp to common/ and
renaming it to reflect that it is used for rocblas types.
