+++ common/include/rocblas_utility.hpp
